Implemented Index Page and Routing
This commit adds the initial routing functionality to the application and sets up the index page to display 'Welcome to the home page!'. Changes include modifications to .htaccess files, updates to the Router class, and adjustments to the index.php file. Additionally, new directories and files related to game views and controllers have been added.